# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: NordiPass Catalog API
  version: "1.0.0"
  description: |
    Protected REST API for the internal NordiPass catalog.
    Provides programmatic access to Categories, Products, Variants,
    Attributes, Media, and lifecycle operations.
    Not intended as a public buyer-facing catalog.
  contact:
    name: NordiPass
servers:
  - url: /api/v1/catalog
    description: Catalog API v1 base
security:
  - bearerAuth: []
paths:
...
  /products/{product}/passport/publish:
    post:
      operationId: publishProductPassport
      summary: Publish the current passport draft
      tags: [Passports]
      security:
        - bearerAuth: [passports.publish]
      parameters:
        - $ref: "#/components/parameters/ProductUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PassportPublishPayload"
      responses:
        "200": { $ref: "#/components/responses/ProductPassportEditor" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /products/{product}/passport/unpublish:
    post:
      operationId: unpublishProductPassport
      summary: Unpublish a published passport
      tags: [Passports]
      security:
        - bearerAuth: [passports.write]
      parameters:
        - $ref: "#/components/parameters/ProductUuid"
      responses:
        "200": { $ref: "#/components/responses/ProductPassportEditor" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /products/{product}/passport/archive:
    post:
      operationId: archiveProductPassport
      summary: Archive an unpublished passport
      tags: [Passports]
      security:
        - bearerAuth: [passports.write]
      parameters:
        - $ref: "#/components/parameters/ProductUuid"
      responses:
        "200": { $ref: "#/components/responses/ProductPassportEditor" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /products/{product}/passport/restore:
    post:
      operationId: restoreProductPassport
      summary: Restore an archived passport to draft
      tags: [Passports]
      security:
        - bearerAuth: [passports.write]
      parameters:
        - $ref: "#/components/parameters/ProductUuid"
      responses:
        "200": { $ref: "#/components/responses/ProductPassportEditor" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
  parameters:
...
    ProductUuid:
      name: product
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object }
        request_id: { type: string, format: uuid }
...
    ErrorResponse:
      type: object
      properties:
        data: { type: "null" }
        meta:
          type: object
          properties:
            request_id: { type: string, format: uuid }
        error:
          $ref: "#/components/schemas/Error"

    # ── Passport schemas ────────────────────────────────
    ProductPassportEditor:
      type: object
      properties:
        passport_uuid: { type: string, format: uuid }
        public_id: { type: string, format: uuid }
        status: { type: string, enum: [draft, published, unpublished, archived] }
        default_language: { type: string }
        enabled_languages: { type: array, items: { type: string } }
        draft_version_uuid: { type: string, format: uuid, nullable: true }
        draft_revision: { type: integer, nullable: true }
        schema_version: { type: string, nullable: true }
        payload: { $ref: "#/components/schemas/DppPayload" }
        created_at: { type: string, format: date-time, nullable: true }
...
    DppPayload:
      type: object
      properties:
        enabled_sections: { type: array, items: { type: string } }
        data: { type: object }
        translations: { type: object }
        document_references:
          type: array
          items:
            $ref: "#/components/schemas/DppDocumentReference"
    DppDocumentReference:
      type: object
      properties:
        document_uuid: { type: string, format: uuid }
        role: { type: string }
        display_order: { type: integer }
...
    PassportPublishPayload:
      type: object
      required: [expected_revision]
      properties:
        expected_revision: { type: integer }
        acknowledge_warnings: { type: boolean, default: false }
...
  responses:
...
    Unauthenticated:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden (ability or permission denied)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found or wrong tenant
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Identifier or state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ProductPassportEditor:
      description: Product passport editor data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProductPassportEditor"
...